Dissociate roles from groups
It may not always be desirable or possible to equate identity management
systems (eg. LDAP) groups with applications roles.
This patch introduces a level of indirection between LDAP groups and
application roles. It is still possible to directly equate LDAP groups
to roles, thanks to the 'groups_to_roles' configuration option. But it
also becomes possible to map LDAP users and/or groups to custom roles
(defined in the configuration file).